Fix compiler [-Wsign-compare] warnings.
+ Tiny code refactoring.

+++ dcpp/AdcHub.cpp
@@ -569,7 +569,7 @@ void AdcHub::handle(AdcCommand::GET, AdcCommand& c) noexcept {
         size_t n = ShareManager::getInstance()->getSharedFiles();
 
         // Ideal size for m is n * k / ln(2), but we allow some slack
-        if(m > (5 * Util::roundUp((int64_t)(n * k / log(2.)), (int64_t)64)) || m > static_cast<size_t>(1 << h)) {
+        if(m > size_t(5 * Util::roundUp((int64_t)(n * k / log(2.)), (int64_t)64)) || m > static_cast<size_t>(1 << h)) {
             send(AdcCommand(AdcCommand::SEV_FATAL, AdcCommand::ERROR_TRANSFER_GENERIC,
                             "Unsupported m", AdcCommand::TYPE_HUB));
             return;
